graph_rag_simplified/requirements.txt
# Core dependencies for Galaxium Document Preprocessing & Vector Indexing
langchain-openai>=0.3.33
langchain-astradb>=0.6.1
python-dotenv>=1.1.1
openai>=1.109.1
astrapy>=1.0.0

# Optional dependencies for enhanced functionality
tqdm>=4.65.0

